Update requirements.txt to allow installation on Windows
Cloning this repo on Windows and trying to `pip install -r requirements.txt` the file leads to a build error with an outdated Cython version. Upgrading Cython to the latest version was however not allowed because of the used pyedbglib having a `< ..` version constraint, so its version was updated to. Thus this tool then installs sucessfully.